refactor(uploads): extract legacy recovery owners
